Drawing analyzer
This program provides a cell for drawing. It has a save option by which you can make the code understand the name for a certain shape you just drew. After, sufficient number of saves, you can draw any object anywhere on the screen of any size from the stuff you have saved previously. The code will tell you the name of the shape drawn on pressing the check button.