Classic Snake Game
This is the remake of the classic snake we all played in the early 2000s. We all loved this game once in our lifetime. It felt great to make this classic game. I always used to think of how complex would it be to make games, but now I realizes it's a piece of cake. It feels great.